incoporated reviews from the merge request

+++ src/expliciteuler.cpp
@@ -27,7 +27,7 @@ void ExplicitEuler::solve(const std::function<double(double, double)>& f)
     double gradient = 0.0;
     _solution[0] = _y_0;
     size_t n = _solution.size();
-    for(int i=1; i<n; i++)
+    for(size_t i=1; i<n; i++)
     {
         if (std::isnan(gradient) || std::isinf(gradient)||std::isnan(_solution.back())||std::isnan(_solution.back()))
         {
